+++ tools/xtask/src/flags.rs
@@ -2,7 +2,17 @@ xflags::xflags! {
     src "./src/flags.rs"
 
     cmd xtask {
-        cmd r {
+        cmd cut-release {
+            /// Cut a patch release
+            optional -p, --patch
+
+            /// Cut a minor release
+            optional -m, --minor
+
+            /// Execute
+            optional --execute
+        }
+        cmd reportgen {
             optional -a, --all
 
             /// Sablier
@@ -51,11 +61,19 @@ pub struct Xtask {
 
 #[derive(Debug)]
 pub enum XtaskCmd {
-    R(R),
+    CutRelease(CutRelease),
+    Reportgen(Reportgen),
 }
 
 #[derive(Debug)]
-pub struct R {
+pub struct CutRelease {
+    pub patch: bool,
+    pub minor: bool,
+    pub execute: bool,
+}
+
+#[derive(Debug)]
+pub struct Reportgen {
     pub all: bool,
     pub sablier: bool,
     pub tg: bool,
